requirements.txt
torch
 torchvision
 transformers
 diffusers
 numpy
 Pillow
 tqdm
